perf: extract lightweight runtime seams
(cherry picked from commit 8727338372)

Discarded files with gutted fork dependencies:
- src/cron/isolated-agent/model-selection.ts (gutted model management)
- src/cron/isolated-agent.model-formatting.test.ts (missing fork deps)
- src/hooks/bundled/session-memory/{handler,handler.test,transcript}.ts (deleted in fork)
