+++ src/hooks/gmail-watcher-errors.ts
@@ -1,0 +1,5 @@
+const ADDRESS_IN_USE_RE = /address already in use|EADDRINUSE/i;
+
+export function isAddressInUseError(line: string): boolean {
+  return ADDRESS_IN_USE_RE.test(line);
+}
